Switched to vector3 from vector3d

+++ GameServer/Config.cs
@@ -8,9 +8,9 @@ namespace GameServer
     {
         public const int BUFFER_SIZE = 4 * 1024;
         public const int SERVER_TICK = 64;
-        public const double PLAYER_SPEED = 2; // PER SECOND
-        public const double GRAVITY = 10; // PER SECOND
-        public const double JUMP_POWER = 2;
-        public const double PLAYER_DECCELERATION = 0.2;
+        public const float PLAYER_SPEED = 2; // PER SECOND
+        public const float GRAVITY = 10; // PER SECOND
+        public const float JUMP_POWER = 2;
+        public const float PLAYER_DECCELERATION = 0.2f;
     }
 }
