+++ GameServer/MapObjects/MapSphere.cs
@@ -7,13 +7,13 @@ namespace GameServer.MapObjects
     public class MapSphere : MapObject
     {
 
-        public MapSphere(double x, double y, double z, double diameter, Color color = null)
+        public MapSphere(float x, float y, float z, float diameter, Color color = null)
             : base(x, y, z, color)
         {
             Diameter = diameter;
         }
 
 
-        public double Diameter { get; set; }
+        public float Diameter { get; set; }
     }
 }
